+++ GWToolbox/GWToolbox/Windows/HotkeysWindow.h
@@ -32,7 +32,7 @@ public:
 	inline bool ToggleRupt() { return ruptActive = !ruptActive; }
 
 	// Update. Will always be called every frame.
-	void Update(DWORD delta) override;
+	void Update(float delta) override;
 
 	// Draw user interface. Will be called every frame if the element is visible
 	void Draw(IDirect3DDevice9* pDevice) override;
